# src/frontend/src/lib/api/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "CurrentUserDto" : {
        "type" : "object",
        "required" : [ "subjectId", "email", "displayName", "roles", "providerName", "language" ],
        "properties" : {
          "subjectId" : {
            "type" : "string"
          },
          "email" : {
            "type" : [ "string", "null" ]
          },
          "displayName" : {
            "type" : [ "string", "null" ]
          },
          "roles" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Role"
            }
          },
          "providerName" : {
            "type" : "string"
          },
          "language" : {
            "type" : "string"
          }
        }
      },
...
      "EstimationCreateDto" : {
        "type" : "object",
        "required" : [ "offer" ],
        "properties" : {
          "offer" : {
            "type" : "string"
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "method" : {
            "$ref" : "#/components/schemas/EstimationMethod"
          }
        }
      },
...
      "EstimationMethod" : {
        "type" : "string",
        "enum" : [ "THREE_POINT_PERT", "BUCKET_SAMPLED_PERT" ]
      },
...
      "EstimationSummaryDto" : {
        "type" : "object",
        "required" : [ "id", "offer", "description", "method", "latestVersionNumber", "versionCount", "hasDraft", "createdAt" ],
        "properties" : {
          "id" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/UUID"
            }, {
              "type" : "null"
            } ]
          },
          "offer" : {
            "type" : [ "string", "null" ]
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "method" : {
            "$ref" : "#/components/schemas/EstimationMethod"
          },
          "latestVersionNumber" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "versionCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "hasDraft" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/Instant"
            }, {
              "type" : "null"
            } ]
          }
        }
      },
...
      "Instant" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T16:15:50Z" ]
      },
      "LanguageUpdateDto" : {
        "type" : "object",
        "required" : [ "language" ],
        "properties" : {
          "language" : {
            "type" : "string"
          }
        }
      },
...
      "ProjectCreateDto" : {
        "type" : "object",
        "required" : [ "name" ],
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "client" : {
            "type" : [ "string", "null" ]
          }
        }
      },
      "ProjectDetailDto" : {
        "type" : "object",
        "required" : [ "id", "name", "description", "client", "status", "createdAt", "estimations" ],
        "properties" : {
          "id" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/UUID"
            }, {
              "type" : "null"
            } ]
          },
          "name" : {
            "type" : [ "string", "null" ]
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "client" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "$ref" : "#/components/schemas/ProjectStatus"
          },
          "createdAt" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/Instant"
            }, {
              "type" : "null"
            } ]
          },
          "estimations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EstimationSummaryDto"
            }
          }
        }
      },
...
      "ProjectStatus" : {
        "type" : "string",
        "enum" : [ "ACTIVE", "ARCHIVED" ]
      },
      "ProjectSummaryDto" : {
        "type" : "object",
        "required" : [ "id", "name", "description", "client", "status", "createdAt" ],
        "properties" : {
          "id" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/UUID"
            }, {
              "type" : "null"
            } ]
          },
          "name" : {
            "type" : [ "string", "null" ]
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "client" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "$ref" : "#/components/schemas/ProjectStatus"
          },
          "createdAt" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/Instant"
            }, {
              "type" : "null"
            } ]
          }
        }
      },
      "ProjectUpdateDto" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : [ "string", "null" ]
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "client" : {
            "type" : [ "string", "null" ]
          }
        }
      },
      "Role" : {
        "type" : "string",
        "enum" : [ "VIEWER", "ESTIMATOR", "ADMIN" ]
      },
...
      "UUID" : {
        "type" : "string",
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      },
...
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "description" : "Authentication",
        "type" : "openIdConnect",
        "openIdConnectUrl" : "https://login.microsoftonline.com/common/v2.0/.well-known/openid-configuration"
      }
    }
  },
...
  "paths" : {
...
    "/api/auth/me" : {
      "get" : {
        "summary" : "The current authenticated user",
        "tags" : [ "Auth" ],
        "parameters" : [ {
          "name" : "Accept-Language",
          "in" : "header",
          "schema" : {
            "type" : [ "string", "null" ]
          },
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "The current user",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CurrentUserDto"
                }
              }
            }
          },
          "401" : {
            "description" : "No user is populated for the request"
          }
        }
      }
    },
    "/api/auth/me/language" : {
      "put" : {
        "summary" : "Update the current user's language preference",
        "tags" : [ "Auth" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LanguageUpdateDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Language preference updated"
          },
          "400" : {
            "description" : "Unsupported language code"
          },
          "401" : {
            "description" : "No user is populated for the request"
          }
        }
      }
    },
...
    "/api/projects" : {
      "get" : {
        "summary" : "List all projects, optionally filtered by status",
        "tags" : [ "Projects" ],
        "parameters" : [ {
          "name" : "status",
          "in" : "query",
          "schema" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/ProjectStatus"
            }, {
              "type" : "null"
            } ]
          },
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "The projects",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProjectSummaryDto"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "VIEWER" ]
        } ]
      },
      "post" : {
        "summary" : "Create a new project",
        "tags" : [ "Projects" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjectCreateDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "The created project",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectSummaryDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/projects/{id}" : {
      "put" : {
        "summary" : "Update a project's metadata",
        "tags" : [ "Projects" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjectUpdateDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "The updated project",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectSummaryDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Project not found"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      },
      "get" : {
        "summary" : "Get a single project with its estimations",
        "tags" : [ "Projects" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The project",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDetailDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Project not found"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "VIEWER" ]
        } ]
      }
    },
    "/api/projects/{id}/archive" : {
      "post" : {
        "summary" : "Archive a project",
        "tags" : [ "Projects" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The archived project",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectSummaryDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Project not found"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/projects/{id}/estimations" : {
      "post" : {
        "summary" : "Create an estimation under a project",
        "tags" : [ "Projects" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EstimationCreateDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "The created estimation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EstimationSummaryDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Project not found"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
...
  },
  "info" : {
    "title" : "Pythia API",
    "version" : "1.0.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}